+++ src/Classes/Apc/ApcCoefficient.php
@@ -44,6 +44,7 @@ class ApcCoefficient
             if ($ressource->hasCoefficientDifferent()) {
                 foreach ($ressource->getSemestres() as $semestre) {
                     $idParcours = $semestre->getDiplome()?->getApcParcours()?->getId();
+                    $tabs['ressources'][$ressource->getId()][$competence->getCompetence()->getId()][$idParcours] = 0;
                     $tabs['ressources'][$ressource->getId()][$idParcours]['total'] = 0;
                 }
             }
@@ -52,7 +53,7 @@ class ApcCoefficient
                     if ($ressource->hasCoefficientDifferent()) {
                         if ($competence->getParcours() !== null) {
                             $tabs['ressources'][$ressource->getId()][$competence->getCompetence()->getId()][$competence->getParcours()->getId()] = $competence->getCoefficient();
-                            $tabs['ressources'][$ressource->getId()][$competence->getParcours()->getId()]['total'] += $competence->getCoefficient();
+                            $tabs['ressources'][$ressource->getId()][$competence->getParcours()?->getId()]['total'] += $competence->getCoefficient();
                         }
 
                         $tabs['competences'][$competence->getCompetence()->getId()]['total'] += $competence->getCoefficient();
@@ -80,6 +81,21 @@ class ApcCoefficient
             $tabs['competences'][$idComp]['total'] = 0;
             $tabs['competences'][$idComp]['totalSaes'] = 0;
             $tabs['competences'][$idComp]['totalRessources'] = 0;
+        }
+
+        return $tabs;
+    }
+
+    private function initIfNotExistParcours(ApcRessourceCompetence $competence, array $tabs)
+    {
+        $idComp = $competence->getCompetence()?->getId();
+        foreach ($this->parcours as $parcours) {
+
+            if (!array_key_exists($idComp, $tabs['competences'])) {
+                $tabs['competences'][$idComp]['total'] = 0;
+                $tabs['competences'][$idComp]['totalSaes'] = 0;
+                $tabs['competences'][$idComp]['totalRessources'] = 0;
+            }
         }
 
         return $tabs;