+++ src/Classes/Edt/MyEdtExport.php
@@ -54,7 +54,7 @@ class MyEdtExport
         KernelInterface                 $kernel
     )
     {
-        $this->dir = $kernel->getProjectDir() . '/public/upload/';
+        $this->dir = $kernel->getProjectDir() . '/public/';
     }
 
     public function export(UserInterface $user, string $_format, string $type, AnneeUniversitaire $anneeUniversitaire): bool|string
@@ -165,7 +165,7 @@ class MyEdtExport
     public function getAllDocs(Departement $departement): array
     {
         // parcour fichiers
-        $folder = $this->dir . 'pdfedt/' . $departement->getId() . '/';
+        $folder = $this->dir . 'upload/pdfedt/' . $departement->getId() . '/';
         $dossier = opendir($folder);
 
         $t = [];
@@ -209,12 +209,18 @@ class MyEdtExport
      */
     public function generePdf(Personnel $personnel, string $source, Departement $departement): void
     {
-        $dir = $this->dir . 'pdfedt/' . $departement->getId() . '/';
-        Tools::checkDirectoryExist($dir);
+        $dir = 'upload/pdfedt/' . $departement->getId();
+        Tools::checkDirectoryExist($this->dir . $dir);
         // todo: passer par le DTO Evenement, comme ca compatible avec celcat
         // todo: gérer l'année universitaire d'export
         if ('intranet' === $source) {
             $planning = $this->edtPlanningRepository->findEdtProf($personnel->getId(), $personnel->getAnneeUniversitaire());
+            //supprimer l'ancien fichier
+            if (file_exists($this->dir . $dir . '/' . $personnel->getId() . '-' . $personnel->getInitiales() . '.pdf')) {
+                unlink($this->dir . $dir . '/' . $personnel->getId() . '-' . $personnel->getInitiales() . '.pdf');
+            }
+
+
             $this->myPDF->pdf()::genereAndSavePdf('pdf/edt/planning.html.twig',
                 [
                     'planning' => $planning,
@@ -222,7 +228,7 @@ class MyEdtExport
                     'departement' => $departement,
                     'matieres' => $this->typeMatiereManager->findByDepartementArray($departement),
                 ],
-                $personnel->getId() . '_' . $personnel->getInitiales(),
+                $personnel->getId() . '-' . $personnel->getInitiales(),
                 $dir);
         }
     }